#59 [feat] 부모 엔티티 삭제 시 자식 엔티티 삭제되도록 매핑

+++ src/main/java/backend/sculptor/domain/museum/dto/MuseumDetailDTO.java
@@ -28,7 +28,7 @@ public final class MuseumDetailDTO {
         private String dDay;
         private Integer powder;
         private Boolean isLike;
-        private Integer achievementRate;
+        private Long achievementRate;
         private Map<String, Long> achievementCounts;
 
         @JsonProperty("dDay")
